+++ cmn-data-svc-client/src/main/java/org/acumos/cds/domain/MLPArtifact.java
@@ -21,7 +21,6 @@
 package org.acumos.cds.domain;
 
 import java.io.Serializable;
-import java.net.URISyntaxException;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -74,11 +73,10 @@ public class MLPArtifact extends MLPAbstractArtifact implements Serializable {
 	 *                             User ID
 	 * @param size
 	 *                             Length
-	 * @throws URISyntaxException
-	 *                                if the URI violates RFC 2396
+	 * @throws IllegalArgumentException
+	 *                                      if the URI value violates RFC 2396
 	 */
-	public MLPArtifact(String version, String artifactTypeCode, String name, String uri, String userId, int size)
-			throws URISyntaxException {
+	public MLPArtifact(String version, String artifactTypeCode, String name, String uri, String userId, int size) {
 		super(version, artifactTypeCode, name, uri, size);
 		if (userId == null)
 			throw new IllegalArgumentException("Null not permitted");