+++ cmn-data-svc-server/src/main/java/org/acumos/cds/controller/SolutionController.java
@@ -46,6 +46,7 @@ import org.acumos.cds.domain.MLPSolutionPicture;
 import org.acumos.cds.domain.MLPSolutionRating;
 import org.acumos.cds.domain.MLPSolutionRating.SolutionRatingPK;
 import org.acumos.cds.domain.MLPSolutionRevision;
+import org.acumos.cds.domain.MLPSolution_;
 import org.acumos.cds.domain.MLPTag;
 import org.acumos.cds.domain.MLPUser;
 import org.acumos.cds.repository.CatSolMapRepository;
@@ -222,17 +223,8 @@ public class SolutionController extends AbstractController {
 	/*
 	 * This method was an early attempt to provide a search feature. Originally
 	 * written with a generic map request parameter to avoid binding field names,
-	 * but that is not supported by Swagger web UI. Now allows use from that web UI
-	 * at the cost of hard-coding many class field names.
+	 * but that is not supported by Swagger web UI. Now allows use from that web UI.
 	 */
-	private static final String NAME = "name";
-	private static final String ACTIVE = "active";
-	private static final String USER_ID = "userId";
-	private static final String SOURCE_ID = "sourceId";
-	private static final String MODEL_TYPE_CODE = "modelTypeCode";
-	private static final String TOOLKIT_TYPE_CODE = "toolkitTypeCode";
-	private static final String ORIGIN = "origin";
-
 	@ApiOperation(value = "Searches for solutions with attributes matching the values specified as query parameters. " //
 			+ "Defaults to match all (conjunction); send junction query parameter '_j=o' to match any (disjunction).", //
 			response = MLPSolution.class, responseContainer = "Page")
@@ -241,20 +233,13 @@ public class SolutionController extends AbstractController {
 	public Object searchSolutions( //
 			@ApiParam(value = "Junction", allowableValues = "a,o") //
 			@RequestParam(name = CCDSConstants.JUNCTION_QUERY_PARAM, required = false) String junction, //
-			@ApiParam(value = "Name") //
-			@RequestParam(name = NAME, required = false) String name, //
-			@ApiParam(value = "Active") //
-			@RequestParam(name = ACTIVE, required = false) Boolean active, //
-			@ApiParam(value = "User ID") //
-			@RequestParam(name = USER_ID, required = false) String userId, //
-			@ApiParam(value = "Source ID") //
-			@RequestParam(name = SOURCE_ID, required = false) String sourceId, //
-			@ApiParam(value = "Model type code") //
-			@RequestParam(name = MODEL_TYPE_CODE, required = false) String modelTypeCode, //
-			@ApiParam(value = "Toolkit type code") //
-			@RequestParam(name = TOOLKIT_TYPE_CODE, required = false) String toolkitTypeCode, //
-			@ApiParam(value = "Origin URI") //
-			@RequestParam(name = ORIGIN, required = false) String origin, //
+			@RequestParam(name = MLPSolution_.NAME, required = false) String name, //
+			@RequestParam(name = MLPSolution_.ACTIVE, required = false) Boolean active, //
+			@RequestParam(name = MLPSolution_.USER_ID, required = false) String userId, //
+			@RequestParam(name = MLPSolution_.SOURCE_ID, required = false) String sourceId, //
+			@RequestParam(name = MLPSolution_.MODEL_TYPE_CODE, required = false) String modelTypeCode, //
+			@RequestParam(name = MLPSolution_.TOOLKIT_TYPE_CODE, required = false) String toolkitTypeCode, //
+			@RequestParam(name = MLPSolution_.ORIGIN, required = false) String origin, //
 			Pageable pageRequest, HttpServletResponse response) {
 		logger.debug("searchSolutions enter");
 		boolean isOr = junction != null && "o".equals(junction);
@@ -440,7 +425,7 @@ public class SolutionController extends AbstractController {
 			return persisted;
 		} catch (Exception ex) {
 			Exception cve = findConstraintViolationException(ex);
-			logger.warn("createSolution failed: {}", cve.toString());
+			logger.warn("createSolution took exception {} on data {}", cve.toString(), solution.toString());
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST, "createSolution failed", cve);
 		}
@@ -473,7 +458,7 @@ public class SolutionController extends AbstractController {
 			return new SuccessTransport(HttpServletResponse.SC_OK, null);
 		} catch (Exception ex) {
 			Exception cve = findConstraintViolationException(ex);
-			logger.warn("updateSolution failed: {}", cve.toString());
+			logger.warn("updateSolution took exception {} on data {}", cve.toString(), solution.toString());
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST, "updateSolution failed", cve);
 		}
@@ -604,7 +589,7 @@ public class SolutionController extends AbstractController {
 			return solutionRevisionRepository.save(revision);
 		} catch (Exception ex) {
 			Exception cve = findConstraintViolationException(ex);
-			logger.warn("createSolutionRevision failed: {}", cve.toString());
+			logger.warn("createSolutionRevision took exception {} on data {}", cve.toString(), revision.toString());
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST, "createSolutionRevision failed", cve);
 		}
@@ -644,7 +629,7 @@ public class SolutionController extends AbstractController {
 			return new SuccessTransport(HttpServletResponse.SC_OK, null);
 		} catch (Exception ex) {
 			Exception cve = findConstraintViolationException(ex);
-			logger.warn("updateSolutionRevision failed: {}", cve.toString());
+			logger.warn("updateSolutionRevision took exception {} on data {}", cve.toString(), revision.toString());
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST, "updateSolutionRevision failed", cve);
 		}
@@ -783,7 +768,7 @@ public class SolutionController extends AbstractController {
 			return result;
 		} catch (Exception ex) {
 			Exception cve = findConstraintViolationException(ex);
-			logger.error("createSolutionDownload failed: {}", cve.toString());
+			logger.warn("createSolutionDownload took exception {} on data {}", cve.toString(), sd.toString());
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST,
 					ex.getCause() != null ? ex.getCause().getMessage() : "createSolutionDownload failed", cve);
@@ -860,7 +845,7 @@ public class SolutionController extends AbstractController {
 			return result;
 		} catch (Exception ex) {
 			Exception cve = findConstraintViolationException(ex);
-			logger.warn("createSolutionRating failed: {}", cve.toString());
+			logger.warn("createSolutionRating took exception {} on data {}", cve.toString(), sr.toString());
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST, "createSolutionRating failed", cve);
 		}
@@ -891,7 +876,7 @@ public class SolutionController extends AbstractController {
 			return new SuccessTransport(HttpServletResponse.SC_OK, null);
 		} catch (Exception ex) {
 			Exception cve = findConstraintViolationException(ex);
-			logger.warn("updateSolutionRating failed: {}", cve.toString());
+			logger.warn("updateSolutionRating took exception {} on data {}", cve.toString(), sr.toString());
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST, "updateSolutionRating failed", cve);
 		}
@@ -1057,7 +1042,7 @@ public class SolutionController extends AbstractController {
 			return result;
 		} catch (Exception ex) {
 			Exception cve = findConstraintViolationException(ex);
-			logger.error("createSolutionDeployment failed: {}", cve.toString());
+			logger.warn("createSolutionDeployment took exception {} on data {}", cve.toString(), sd.toString());
 			response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
 			return new ErrorTransport(HttpServletResponse.SC_INTERNAL_SERVER_ERROR,
 					ex.getCause() != null ? ex.getCause().getMessage() : "createSolutionDeployment failed", cve);
@@ -1090,7 +1075,7 @@ public class SolutionController extends AbstractController {
 		} catch (Exception ex) {
 			// e.g., EmptyResultDataAccessException is NOT an internal server error
 			Exception cve = findConstraintViolationException(ex);
-			logger.warn("updateSolutionDeployment failed: {}", cve.toString());
+			logger.warn("updateSolutionDeployment took exception {} on data {}", cve.toString(), sd.toString());
 			response.setStatus(HttpServletResponse.SC_BAD_REQUEST);
 			return new ErrorTransport(HttpServletResponse.SC_BAD_REQUEST, "updateSolutionDeployment failed", cve);
 		}